New dictionary-based interface based on ogg.vorbis, initial changes by Martin Michlmayr.
Default genre is set to 255, not 0, with new ID3 objects.

New as_tuple argument to __init__, to preserve compatibility with
ogg.vorbis (returns all dictionary values as tuples).  Defaults to 0,
since I think most people won't need this, but we want to be interface-
compatible with ogg.vorbis for the people who need it.